#5765 Move logging level names
